% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeriToolsV12.R
\name{computeAllNBVPC}
\alias{computeAllNBVPC}
\title{Calculate the negative binomial vpc for one or more features.}
\usage{
computeAllNBVPC(para)
}
\arguments{
\item{para}{A \eqn{1 \times 3}{1x3} vector or \eqn{k \times 3}{k x 3} matrix of negative binomial fit parameters.
Order of the parameters: \eqn{({\beta}_0, {\sigma_a^2, \phi)}}{(beta0, sigma_a2, phi)}.}
}
\value{
A \eqn{G \times 1}{G x 1} matrix consisting of variance partition coefficients for
\eqn{G} features based on negative binomial mixed model (NBMM). Column name is "NBMM_vpc"; row names are the 
feature names.
}
\description{
Calculate the negative binomial vpc for one or more features.
}

