% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeriToolsV12.R
\name{fitandcompute_lmer_VPC}
\alias{fitandcompute_lmer_VPC}
\title{Fit linear mixed models (LMM) and compute the VPC values for one or more 
features.}
\usage{
fitandcompute_lmer_VPC(CountMatrix, Strains, PriorWeights = NULL,
  test = FALSE, VPCname = "LMM")
}
\arguments{
\item{CountMatrix}{Sequencing count matrix for one or more features. Each 
row is for one feature, and the columns are for samples.}

\item{Strains}{Strain labels for the samples.}

\item{PriorWeights}{Weights used in the lmer function in the package lme4. 
It is an optional vector used in the fitting process.}

\item{test}{TRUE or FALSE (default). Test the presence of heritability 
through examining \eqn{\sigma_g^2 = 0}{}}

\item{VPCname}{Name of the VPC result, default = "LMM".}
}
\value{
A list with two objects. The first object is a 
\eqn{1 \times G}{1 x G} vector indicating the variance partition coefficients
(VPC). If the argument test is set to be TRUE, the second object consists 
of the p-values from testing the presence of random effect. If test = FALSE, 
the second object is NULL.
}
\description{
Fit the Gaussian-like data to LMM and compute the VPC values for 
one or more features.
}
\examples{
## Compute VPC for each feature under linear mixed models for Gaussian-like datasets. 

## Provide normalized data and include hypothesis testing on presence of
## heritability:
result.vst <- fitandcompute_lmer_VPC(simData_vst, strains, test = TRUE)
## Extract parameters
vpc.vst <- result.vst[[1]]
## Extract p-values
pval.vst <- result.vst[[2]]

## Visulize the distribution of p-values.
hist(pval.vst, breaks = 30, col = "cyan")
}

