% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeriToolsV12.R
\name{fitCPMM}
\alias{fitCPMM}
\title{Fit a compound Poisson mixed effect model for a list of features}
\usage{
fitCPMM(CountMatrix, Strains, test = FALSE, optimizer = "nlminb")
}
\arguments{
\item{CountMatrix}{Sequencing count matrix for a list of features. Each row 
is for one feature, and the columns are for samples.}

\item{Strains}{Strain labels for the samples.}

\item{test}{TRUE or FALSE (default). Test the presence of heritability through examining \eqn{\sigma_a^2 = 0}{}}

\item{optimizer}{A character string that determines which optimization routine is}
}
\value{
A list with two members. The first is a \eqn{G \times 4}{G x 4} matrix 
indicating the fitted parameters for each feature The columns are ordered by 
"beta0", "p", sigma_a2", "phi". Row names are feature names; the second member of the list  
consists of p-values for testing the hypothesis that \eqn{\sigma_a^2 = 0}{sigma_a2 = 0}.
}
\description{
Fit a CPMM for a list of features.
}

