% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeriToolsV12.R
\name{fitNBMM}
\alias{fitNBMM}
\title{Fit a NBMM for a list of features}
\usage{
fitNBMM(CountMatrix, Strains, test = FALSE)
}
\arguments{
\item{CountMatrix}{Sequencing count matrix for a list of features. Each row 
is for one feature, and the columns are for samples.}

\item{Strains}{Strain labels for the samples.}

\item{test}{TRUE or FALSE (default). Test the presence of heritability through examining \eqn{\sigma_a^2 = 0}{}}
}
\value{
A list with two members. The first is a \eqn{G \times 3}{G x 3} matrix indicating the fitted parameters 
for each feature The columns are ordered by "beta0", "sigma_a2", "phi". 
Row names are feature names; the second member of the list  
consists of p-values for testing the hypothesis that \eqn{\sigma_a^2 = 0}{sigma_a2 = 0}.
}
\description{
Fit a NBMM for a list of features and output the fit parameters.
}
\examples{
## Compute vpc for each feature under NBMM. This will take a while on the
##  entire dataset. For the purpose of illustration, here we only fit on 
##  the first 10 features.
result.nb <- fitNBMM(simData[1:10, ], strains)
}

