% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{para_cp}
\alias{para_cp}
\title{Parameter matrix obtained from simData by fitting CPMM.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 4 columns.
}
\usage{
para_cp
}
\description{
Parameter matrix obtained from simData by fitting CPMM.
}
\keyword{datasets}
