% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{para_nb}
\alias{para_nb}
\title{Parameter matrix obtained from simData by fitting NBMM.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 3 columns.
}
\usage{
para_nb
}
\description{
Parameter matrix obtained from simData by fitting NBMM.
}
\keyword{datasets}
