% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simData}
\alias{simData}
\title{A simulated sequencing dataset.}
\format{
A matrix with 100 rows and 175 columns
}
\usage{
simData
}
\description{
A matrix containing simulated counts for 100 features (rows) and 175 samples (columns)
}
\keyword{datasets}
