% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simData_voom}
\alias{simData_voom}
\title{Voom transformed version of simData.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 881 rows and 175 columns.
}
\usage{
simData_voom
}
\description{
Voom transformed version of simData.
}
\keyword{datasets}
