% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simData_vst}
\alias{simData_vst}
\title{Variance stabilize transformed version of simData.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 881 rows and 175 columns.
}
\usage{
simData_vst
}
\description{
Variance stabilize transformed version of simData.
}
\keyword{datasets}
