% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strains}
\alias{strains}
\title{List of strain names for the samples.}
\format{
An object of class \code{character} of length 175.
}
\usage{
strains
}
\description{
List of strain names for the samples.
}
\keyword{datasets}
