% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weights_voom}
\alias{weights_voom}
\title{Weights used in the voom transformation.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 881 rows and 175 columns.
}
\usage{
weights_voom
}
\description{
Weights used in the voom transformation.
}
\keyword{datasets}
