% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeriToolsV12.R
\name{computeAllCPVPC}
\alias{computeAllCPVPC}
\title{Calculate the compound Poisson VPC for one or more features.}
\usage{
computeAllCPVPC(para)
}
\arguments{
\item{para}{A \eqn{1 \times 4}{1 x 4} vector or \eqn{k \times 4}{k x 4} matrix of negative binomial fit parameters.
Order of the parameters: \eqn{({\beta}_0, {\sigma_a^2, p, \phi)}}{(beta0, sigma_a2, p, phi)}.}
}
\value{
vpcs A \eqn{G \times 1}{G x 1} matrix consisting of variance partition coefficients for
  G features based on compound Poisson mixture model (CPMM) Column name is "CPMM_vpc"; row names are the 
  feature names.
}
\description{
Calculate the compound Poisson VPC for one or more features.
}

