% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeriToolsV13.R
\name{computeAlllmerVPC}
\alias{computeAlllmerVPC}
\title{Compute the VPC values for a list of features under a linear mixed model (LMM).}
\usage{
computeAlllmerVPC(CountMatrix, Strains, PriorWeights = NULL, test = FALSE)
}
\arguments{
\item{CountMatrix}{Sequencing count matrix for a list of features. Each row 
is for one feature, and the columns are for samples.}

\item{Strains}{Strain labels for the samples.}

\item{PriorWeights}{Weights used in the lmer function.}

\item{test}{TRUE or FALSE (default). Test the presence of heritability through examining \eqn{\sigma_a^2 = 0}{}}
}
\value{
A list with two members. The first is a \eqn{1 \times G}{1 x G} vector indicating the significance or random effects
 for each feature under linear mixed model (LMM); the second member of the list consists of the 
 p-values from testing the hypothesis that there is no random effect.
}
\description{
Compute the VPC values for a list of features under a linear mixed model (LMM).
}
\examples{
## Compute vpc for each feature under LMM.

## Provide normalized data with prior weights:
result.voom <- computeAlllmerVPC(simData_voom, strains, PriorWeights = weights_voom)
vpc.voom <- result.voom[[1]]

## Provide normalized data without prior weights and include hypothesis 
##  testing on presence of heritability:
result.vst <- computeAlllmerVPC(simData_vst, strains, test = TRUE)
## Extract parameters
vpc.vst <- result.vst[[1]]
## Extract p-values
pval.vst <- result.vst[[2]]

## Visulize the distribution of p-values.
hist(pval.vst, breaks = 30, col = "cyan")
}

