% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeriToolsV12.R
\name{fitCPMM}
\alias{fitCPMM}
\title{Fit a compound Poisson mixed effect model for a list of features}
\usage{
fitCPMM(CountMatrix, Strains, test = FALSE, optimizer = "nlminb")
}
\arguments{
\item{CountMatrix}{Sequencing count matrix for a list of features. Each row 
is for one feature, and the columns are for samples.}

\item{Strains}{Strain labels for the samples.}

\item{test}{TRUE or FALSE (default). Test the presence of heritability through examining \eqn{\sigma_a^2 = 0}{}}

\item{optimizer}{A character string that determines which optimization routine is}
}
\value{
A list with two members. The first is a \eqn{G \times 4}{G x 4} matrix 
indicating the fitted parameters for each feature The columns are ordered by 
"beta0", "p", sigma_a2", "phi". Row names are feature names; the second member of the list  
consists of p-values for testing the hypothesis that \eqn{\sigma_a^2 = 0}{sigma_a2 = 0}.
## Fit CPMM on the entire dataset takes about 30 minutes. For the purpose 
##  of illustration, here we only fit on the first 10 features.
}
\description{
Fit a CPMM for one or more features.
}
\examples{
\donttest{
## Fit CPMM for the first two features using the default optimizer. 
result.cp <- fitCPMM(simData[1, ], strains)
## Extract parameters
para.cp <- result.cp[[1]]
## Extract p-values
pval.cp <- result.cp[[2]]

## Fit CPMM for the first two features using an alternative optimizer.
result.cp2 <- fitCPMM(simData[1:2, ], strains, optimizer = "L-BFGS-B")

## Fit CPMM for the first two features with hypothesis testing 
##  on presence of heritability.
result.cp3 <- fitCPMM(simData[1:2, ], strains, test = TRUE)
}
}

