% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeriToolsV13.R
\name{getNBReadMatrix}
\alias{getNBReadMatrix}
\title{Generate a count matrix from NBMM models.}
\usage{
getNBReadMatrix(vec.num.rep, beta0s, sig2.strains, phis)
}
\arguments{
\item{vec.num.rep}{A vector of replicate numbers for each strain.}

\item{beta0s}{Vector for the intercepts in the GLMM's.}

\item{sig2.strains}{Variance vector of the strain random effect.}

\item{phis}{Dispersion vector in the NB models; common across strains per feature.}
}
\value{
A \eqn{G \times N}{G x N} matrix with NB counts. \eqn{N} is the total number of 
 samples; \eqn{G} is the number of features. Column names are sample names
 of the form "Ss_r", where S stands for sample, s is the strain number,
 r is the replicate number within the strain. Row names are the feature
 names of the form "Gene g", where g is the feature index.
}
\description{
Generate a (possibly unbalanced) count matrix from
negative binomial mixed effect models.
Each row corresponds to one feature that follows: \cr
\eqn{log(E(Y_{ij})) = {\beta}_0 + Strain_i}{log(E(Yij)) = beta0 + Strain_i} \cr
\eqn{Strain_i \sim N(0, sig2.strain)}{Strain_i ~ N(0, sig2.strain)}
}
\examples{
## Generate a sequencing dataset with 5 features and 6 strains. 
## Assign parameter values.
rep.num <- c(3, 5, 2, 3, 4, 2)
b0 <- c(-1, 1, 2, 5, 10)
sig2s <- c(10, 0.2, 0.1, 0.03, 0.01)
phis <- c(0.5, 1, 0.05, 0.01, 0.1)

set.seed(1234)
## Generate reads:
nbData <- getNBReadMatrix(rep.num, b0, sig2s, phis)
}

